hi5/requirements.live.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.24.0
yfinance>=0.2.28

# Google Cloud Storage
google-cloud-storage>=2.14.0

# Email and Calendar
icalendar>=5.0.0
pytz>=2023.3

# Date and Time
python-dateutil>=2.8.2

# Optional: Backtrader (if needed for backtesting)
backtrader>=1.9.76.123

# Additional utilities
asyncio>=3.4.3

# Google API dependencies
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
google-auth>=2.23.0 
google-cloud-bigquery>=3.11.0
db-dtypes>=1.4.3

openpyxl>=3.1.5
jinja2>=3.1.2
